chore(example): Rename files to Dart style (#2618)
* 1. rename LandingPage.dart as landing_page.dart

* renamed LoadingPage.dart as loading_page.dart

* renamed MainPage.dart as main_page.dart

* renamed ErrorView.dart as error_view.dart

* renamed ImageLineItem.dart as image_line_item.dart

* renamed ImagePreview.dart as image_preview.dart

* renamed ImageUploader.dart as image_uploader.dart

* renamed SignInView.dart as sign_in_view.dart

* renamed SignUpView.dart as sign_up_view.dart

* renamed UserView.dart as user_view.dart

* updated imports in main.dart

* renamed ModelProvider.dart as model_provider.dart

* updated imports in main.dart

* updated imports wrt to folder names

* Rename ModelProvider file to match `amplify codegen models` output

---------
