+++ example/lib/main.dart
@@ -18,9 +18,8 @@ import 'package:amplify_auth_cognito/amplify_auth_cognito.dart';
 import 'package:amplify_flutter/amplify_flutter.dart';
 import 'package:amplify_storage_s3/amplify_storage_s3.dart';
 import 'package:flutter/material.dart';
-import 'package:sample_app/Pages/LoadingPage.dart';
-
-import 'Pages/LandingPage.dart';
+import 'package:sample_app/pages/landing_page.dart';
+import 'package:sample_app/pages/loading_page.dart';
 
 /*
 IMPORTANT - THIS LINE WILL NOT COMPILE
