+++ example/lib/pages/loading_page.dart
