+++ example/lib/views/image_line_item.dart
@@ -15,7 +15,7 @@
 
 import 'package:flutter/material.dart';
 
-import 'ImagePreview.dart';
+import 'image_preview.dart';
 
 class ImageLineItem extends StatelessWidget {
   final String storageKey;
